Move UACutoffWarning to common folder.
